refactor SidebarTab to new component
